Rename component files to match function names
