+++ src/app/edit/AppDetails.tsx
